use react toast in student list
